third_party/blink/web_tests/virtual/customizable-select-disabled/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-parsing.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] <div>s, <button>s, and <datalist>s should be allowed in <select>.
  assert_equals: expected "<div>div 1</div>\\n  <button>button</button>\\n  <div>div 2</div>\\n  <datalist>\\n    <option>option</option>\\n  </datalist>\\n  <div>div 3</div>" but got "div 1\\n  button\\n  div 2\\n  \\n    <option>option</option>\\n  \\n  div 3"
[FAIL] </select> should close <button>.
  assert_equals: expected "<button>button\\n</button>" but got "button"
[FAIL] </select> should close <datalist>.
  assert_equals: expected "<datalist>datalist\\n</datalist>" but got "datalist"
[FAIL] <select> in <button> in <select> should remove inner <select>.
  assert_equals: expected "<button>\\n    </button>" but got ""
[FAIL] <select> in <select><button><div> should remove inner <select>.
  assert_equals: expected "<button>\\n    <div>\\n      </div></button>" but got ""
[FAIL] Divs and imgs should be allowed as direct children of select and within options without a datalist.
  assert_equals: expected "<div>\\n    <option><img>option</option>\\n  </div>" but got "<option>option</option>"
[FAIL] Input tags should parse inside select instead of closing the select.
  assert_equals: expected "<input>" but got ""
Harness: the test ran to completion.
